chore(analyze): move analyze to pkg
